Remove all permissions
Removes all permissions from the extension.  This precludes the usage of the "safe method," which is also removed here.

Rationale:

The "safe method" is an option to make a separate HTTP request to get the raw HTTP content.

It requires granting the extension permission to perform requests to arbitrary URLs.

I suspect granting these permissions for a feature that isn't commonly used actually poses the greater risk.